fixed addBudget screen to be consistent with addEntry screen

+++ src/com/example/budgetmanager/AddBudgetActivity.java
@@ -100,7 +100,7 @@ public class AddBudgetActivity extends Activity {
 		}
 		
 		// checks whether the amount is non-zero
-		if (Integer.parseInt(mBudgetAmountView.getText().toString()) == 0) {
+		if (Double.parseDouble(mBudgetAmountView.getText().toString()) == 0.0) {
 			mBudgetAmountView.setError(getString(R.string.error_zero_amount));
 			mBudgetAmountView.requestFocus();
 			return;
